+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/Icons.kt
@@ -26,8 +26,6 @@ private fun getIcon(path: String): Icon = IconLoader.getIcon(path, WavefrontObjB
 
 val MTL_FILE_ICON: Icon = getIcon("/icons/mtlFileIcon.svg")
 
-val MTL_ERROR_ICON: Icon = getIcon("/icons/mtlErrorIcon.svg")
-
 val MTL_MATERIAL_ICON: Icon = getIcon("/icons/mtlMaterialIcon.svg")
 
 val MTL_PROPERTY_ICON: Icon = getIcon("/icons/mtlPropertyIcon.svg")
@@ -35,8 +33,6 @@ val MTL_TEXTURE_ICON: Icon = getIcon("/icons/mtlTextureIcon.svg")
 val MTL_OPTION_ICON: Icon = getIcon("/icons/mtlOptionIcon.svg")
 
 val OBJ_FILE_ICON: Icon = getIcon("/icons/objFileIcon.svg")
-
-val OBJ_ERROR_ICON: Icon = getIcon("/icons/objErrorIcon.svg")
 
 val OBJ_OBJECT_ICON: Icon = getIcon("/icons/objObjectIcon.svg")
 val OBJ_GROUP_ICON: Icon = getIcon("/icons/objGroupIcon.svg")